# openapi/models/models.gen.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package models

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Item defines model for Item.
type Item struct {
	// Attribute アイテムの属性 (1~6の整数)
	Attribute int `json:"attribute"`

	// Description アイテムのフレーバーテキスト
	Description string `json:"description"`

	// Id アイテムID
	Id openapi_types.UUID `json:"id"`

	// Name アイテムの名前
	Name string `json:"name"`

	// Rarity 強さ度合い(レア度) (0~1の数字)
	Rarity int `json:"rarity"`
}

// CombineItemsJSONBody defines parameters for CombineItems.
type CombineItemsJSONBody struct {
	// ItemId1 合成する最初のアイテムID
	ItemId1 *openapi_types.UUID `json:"itemId1,omitempty"`

	// ItemId2 合成する二番目のアイテムID
	ItemId2 *openapi_types.UUID `json:"itemId2,omitempty"`
}

// GetItemWithParametersJSONBody defines parameters for GetItemWithParameters.
type GetItemWithParametersJSONBody struct {
	// ItemId アイテムID
	ItemId     *openapi_types.UUID `json:"itemId,omitempty"`
	Parameters *struct {
		// Attack 攻撃力
		Attack *int `json:"Attack,omitempty"`

		// Defence 防御力
		Defence *int `json:"Defence,omitempty"`

		// HP ヒットポイント
		HP *int `json:"HP,omitempty"`
	} `json:"parameters,omitempty"`
}

// CombineItemsJSONRequestBody defines body for CombineItems for application/json ContentType.
type CombineItemsJSONRequestBody CombineItemsJSONBody

// GetItemWithParametersJSONRequestBody defines body for GetItemWithParameters for application/json ContentType.
type GetItemWithParametersJSONRequestBody GetItemWithParametersJSONBody